:hammer: Refactor data models directory
