+++ app/src/main/java/com/rummenigged/popcorntime/data/model/safe/SeasonSafe.kt
@@ -1,4 +1,4 @@
-package com.rummenigged.popcorntime.data.model
+package com.rummenigged.popcorntime.data.model.safe
 
 import com.rummenigged.popcorntime.data.common.Safe
 import com.rummenigged.popcorntime.domain.Season
